boot: move bootloaderKernelState20 impls to separate file
Purely aesthetic to keep bootstate20.go smaller/clearer.
